+++ user/phone/disable-teams-phone.ps1
@@ -4,12 +4,12 @@
 
   .DESCRIPTION
   Remove the phone number and specific policies from a teams-enabled user.
-  Note: A Microsoft Teams service account must be available and stored - details can be found in the runbook.
   
   .NOTES
   Permissions: 
   The connection of the Microsoft Teams PowerShell module is ideally done through the Managed Identity of the Automation account of RealmJoin.
-  If this has not yet been set up and the old "Service User" is still stored, the connect is still included for stability reasons. However, it should be switched to Managed Identity as soon as possible.
+  If this has not yet been set up and the old "Service User" is still stored, the connect is still included for stability reasons. 
+  However, it should be switched to Managed Identity as soon as possible!
 
   .INPUTS
   RunbookCustomization: {
@@ -21,7 +21,7 @@
 }
 #>
 
-#Requires -Modules @{ModuleName = "RealmJoin.RunbookHelper"; ModuleVersion = "0.8.3" }, @{ModuleName = "MicrosoftTeams"; ModuleVersion = "5.8.0" }
+#Requires -Modules @{ModuleName = "RealmJoin.RunbookHelper"; ModuleVersion = "0.8.3" }, @{ModuleName = "MicrosoftTeams"; ModuleVersion = "5.9.0" }
 
 param(
     [Parameter(Mandatory = $true)]